requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
email-validator==2.2.0
qrcode==7.4.2
Pillow==10.3.0
reportlab==4.2.2
